Switch to using serviceId as Kafka key
Going to try to use Kafka keys to fan in/fan out messages on the
message-input and message-output channels.  Alternative is creating a
single service that does this, but seems like extra work for little gain
unless the number of messages get crazy high.

And in that case, the answer is likely switching to Kubernetes and
spinning up more services in a consumer group anyway.